+++ src/config/mod.rs
@@ -12,9 +12,9 @@ pub const MAX_DELTA_SIZE: usize = 1024 * 1024 * 1024; // Bytes
 
 // Every n blocks, recalculate difficulty
 pub const BLOCK_TIME: usize = 60; // Seconds
-pub const DIFFICULTY_CALC_INTERVAL: usize = 128; // Blocks
+pub const DIFFICULTY_CALC_INTERVAL: u64 = 128; // Blocks
 
-pub const MAX_BLOCK_FETCH: usize = 16; // Blocks
+pub const MAX_BLOCK_FETCH: u64 = 16; // Blocks
 
 // 0 63 -> BAZUKA BASE KEY
 // 64 2111 -> hash(blk#0)
@@ -22,12 +22,12 @@ pub const MAX_BLOCK_FETCH: usize = 16; // Blocks
 // 4160 6207 -> hash(blk#4096)
 // ...
 pub const POW_BASE_KEY: &[u8] = b"BAZUKA BASE KEY";
-pub const POW_KEY_CHANGE_DELAY: usize = 64; // Blocks
-pub const POW_KEY_CHANGE_INTERVAL: usize = 2048; // Blocks
+pub const POW_KEY_CHANGE_DELAY: u64 = 64; // Blocks
+pub const POW_KEY_CHANGE_INTERVAL: u64 = 2048; // Blocks
 
 // New block's timestamp should be higher than median
 // timestamp of 10 previous blocks
-pub const MEDIAN_TIMESTAMP_COUNT: usize = 10;
+pub const MEDIAN_TIMESTAMP_COUNT: u64 = 10;
 
 // Our Zero-Knowledge RAM will have 2^32 memory cells
 pub const LOG_ZK_RAM_SIZE: usize = 32;
