+++ src/node/api/post_miner_solution.rs
@@ -23,7 +23,7 @@ pub async fn post_miner_solution<B: Blockchain>(
     block.header.proof_of_work.nonce = u64::from_le_bytes(nonce_bytes);
     if context
         .blockchain
-        .extend(block.header.number as usize, &[block])
+        .extend(block.header.number, &[block])
         .is_ok()
     {
         context.miner.as_mut().unwrap().block_puzzle = None;
